create card questions  component
