+++ client/src/components/QuestionsPage/CardsQuestion/style.css
@@ -1,0 +1,23 @@
+.box {
+  list-style-type: none;
+}
+
+.box__card {
+  width: auto;
+  height: 1.5rem;
+  margin-left: 0.7rem;
+  margin-right: 0.7rem;
+  margin-bottom: 1rem;  background-color: white;
+  padding: 2rem 2rem 3rem 2rem;
+  box-shadow: 0px 0px 23px -5px rgba(0, 0, 0, 0.75);
+}
+.box__card--content{
+  display: flex;
+  justify-content: space-between;
+  color: rgba(0, 0, 0, 0.38);
+}
+.box__card--content--box{
+  display: grid;
+  grid-template-columns: 1fr 1fr;
+  grid-gap: 10px;
+}
