--- client/src/components/common/Cards/style.css
@@ -1,13 +1,0 @@
-.box {
-  list-style-type: none;
-}
-
-.box__card {
-  width: 46rem;
-  height: 1.5rem;
-  margin-left: 0.7rem;
-  margin-right: 0.7rem;
-  margin-bottom: 1rem;  background-color: white;
-  padding: 2rem;
-  box-shadow: 0px 0px 23px -5px rgba(0, 0, 0, 0.75);
-}
